Create plotly_gauge.ipynb
Testing plotly gauge for humidity readings. Would be pulled from Weather Station API.  Line 4 'wxdata' is a stand in for the array value that would be brought in from that API. See different paths under shape for examples of how the arrow should be changed based on percentage. Will need to make a variable equation to figure out the correct path.